Fix iOS implementation
The kubenav app requires a KubenavKubenavPlugin.podspec instead of the
KubenavPlugin.podspec file. Maybe this is to the fact that we install
the NPM package und @kubenav/kubenav-plugin, but I'm not sure how this
is related.

When we publish the NPM package the symlinks in the Request.framework
file get lost. So we removed all symlinks and placed the Headers,
Modules and Resources folder and the Request file in the root of the
Request.framework folder.